color picker working conditon Happy new Year
