+++ public/index.html
@@ -20,6 +20,8 @@
   <link rel="stylesheet" href="%PUBLIC_URL%/assets/css/main.css" />
   <link rel="stylesheet" href="%PUBLIC_URL%/assets/css/customizer.css" />
   <link rel="stylesheet" href="%PUBLIC_URL%/assets/css/custom.css" />
+  <!-- <link rel="stylesheet" href="%PUBLIC_URL%/assets/css/custom-color.css" /> -->
+  
 
 
   <!--
@@ -31,10 +33,36 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-
+    
   <title>
     HRMGo - Dashboard - Nikita Testing
   </title>
+
+   <script>
+    // Function to fetch the business settings
+    async function fetchBusinessSettings() {
+      try {
+        // Replace with your actual API endpoint
+        const response = await fetch('/get-business-setting');
+        
+        if (!response.ok) {
+          throw new Error('Failed to fetch business settings');
+        }
+
+        const { data } = await response.json();
+
+        // Update the document's title dynamically
+        if (data && data.titleText) {
+          document.title = data.titleText; // Use the titleText from the API
+        }
+      } catch (error) {
+        console.error('Error fetching business settings:', error.message);
+      }
+    }
+
+    // Call the function on page load
+    fetchBusinessSettings();
+  </script>
 
   <!-- SEO META -->
   <meta name="title" content="HRMGo">
@@ -68,8 +96,8 @@
 
 </head>
 
-<body class="theme-3">
-  <noscript>You need to enable JavaScript to run this app.</noscript>
+<body class="">
+  <!-- <noscript>You need to enable JavaScript to run this app.</noscript> -->
   <div id="root"></div>
   <!--
       This HTML file is a template.
